# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Muhandis Islamic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type QuranSummaryLastRead = {
  surahId?: number;
  surahName?: string;
  ayahNumber?: number;
} | null;

export interface QuranSummary {
  totalSurahs: number;
  totalAyahs: number;
  totalWords: number;
  makkiSurahs: number;
  madaniSurahs: number;
  lastRead?: QuranSummaryLastRead;
}

export type SurahRevelation =
  (typeof SurahRevelation)[keyof typeof SurahRevelation];

export const SurahRevelation = {
  Meccan: "Meccan",
  Medinan: "Medinan",
} as const;

export interface Surah {
  id: number;
  number: number;
  nameArabic: string;
  nameTransliterated: string;
  nameEnglish: string;
  revelation: SurahRevelation;
  ayahCount: number;
  juzNumber: number;
  description?: string;
}

export interface Ayah {
  id: number;
  surahId: number;
  surahName: string;
  ayahNumber: number;
  arabicText: string;
  translation: string;
  transliteration: string;
  juzNumber: number;
  pageNumber: number;
  audioUrl?: string | null;
}

export interface AyahWord {
  id: number;
  ayahId: number;
  position: number;
  arabicText: string;
  transliteration: string;
  translation: string;
  rootWord: string;
  grammar: string;
}

export interface Tafseer {
  ayahId: number;
  source: string;
  arabicText: string;
  englishText: string;
  scholarName: string;
}

export interface Translation {
  id: string;
  name: string;
  language: string;
  author: string;
}

export interface Reciter {
  id: string;
  name: string;
  style: string;
  language: string;
}

export type HadithSummaryGradeBreakdown = {
  sahih: number;
  hasan: number;
  daif: number;
  munkar?: number;
  mawdu?: number;
  unknown?: number;
};

export interface HadithSummary {
  totalCollections: number;
  totalHadiths: number;
  totalNarrators: number;
  gradeBreakdown: HadithSummaryGradeBreakdown;
}

export interface HadithCollection {
  id: string;
  name: string;
  nameArabic: string;
  author: string;
  totalHadiths: number;
  description: string;
  era: string;
}

export type HadithGrade = (typeof HadithGrade)[keyof typeof HadithGrade];

export const HadithGrade = {
  Sahih: "Sahih",
  Hasan: "Hasan",
  "Da'if": "Da'if",
  "Mawdu'": "Mawdu'",
  Unknown: "Unknown",
} as const;

export interface Hadith {
  id: number;
  collectionId: string;
  collectionName: string;
  hadithNumber: string;
  arabicText: string;
  translation: string;
  grade: HadithGrade;
  gradeReason?: string;
  narrator: string;
  topics: string[];
}

export interface Narrator {
  id: number;
  name: string;
  nameArabic: string;
  birthYear?: number | null;
  deathYear?: number | null;
  location?: string;
  reliability: string;
  grade: string;
  heardFrom: string[];
  position: number;
}

export interface IsnadChain {
  hadithId: number;
  narrators: Narrator[];
  overallGrade: string;
}

export type HadithDetail = Hadith & {
  sharh?: string;
  isnadChain?: IsnadChain;
};

export interface HadithListResponse {
  hadiths: Hadith[];
  total: number;
  page: number;
  limit: number;
}

export interface SearchResults {
  query: string;
  ayahs: Ayah[];
  hadiths: Hadith[];
  total: number;
}

export type BookmarkType = (typeof BookmarkType)[keyof typeof BookmarkType];

export const BookmarkType = {
  ayah: "ayah",
  hadith: "hadith",
} as const;

export interface Bookmark {
  id: number;
  type: BookmarkType;
  referenceId: number;
  title: string;
  note?: string | null;
  createdAt: string;
}

export type CreateBookmarkBodyType =
  (typeof CreateBookmarkBodyType)[keyof typeof CreateBookmarkBodyType];

export const CreateBookmarkBodyType = {
  ayah: "ayah",
  hadith: "hadith",
} as const;

export interface CreateBookmarkBody {
  type: CreateBookmarkBodyType;
  referenceId: number;
  title: string;
  note?: string | null;
}

export type ListAyahsParams = {
  translationId?: string;
};

export type GetAyahTafseerParams = {
  source?: string;
};

export type ListHadithsParams = {
  page?: number;
  limit?: number;
};

export type SearchParams = {
  q: string;
  type?: SearchType;
};

export type SearchType = (typeof SearchType)[keyof typeof SearchType];

export const SearchType = {
  all: "all",
  quran: "quran",
  hadith: "hadith",
} as const;