# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Al-Muhandis Islamic Intelligence Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /hadith/collections/{collectionId}/hadiths:
    get:
      operationId: listHadiths
      tags: [hadith]
      summary: List hadiths in a collection
      parameters:
        - name: collectionId
          in: path
          required: true
          schema:
            type: string
        - name: page
          in: query
          required: false
          schema:
            type: integer
        - name: limit
          in: query
          required: false
          schema:
            type: integer
      responses:
        "200":
          description: Hadiths list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HadithListResponse"
  /hadith/{hadithId}:
    get:
      operationId: getHadith
      tags: [hadith]
      summary: Get a specific hadith with full details
      parameters:
        - name: hadithId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Hadith details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HadithDetail"
  /hadith/{hadithId}/isnad:
    get:
      operationId: getHadithIsnad
      tags: [hadith]
      summary: Get isnad chain for a hadith
      parameters:
        - name: hadithId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Isnad chain
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IsnadChain"
  /hadith/{hadithId}/related-ayahs:
    get:
      operationId: getHadithRelatedAyahs
      tags: [hadith]
      summary: Get Quranic ayahs related to a hadith
      parameters:
        - name: hadithId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Related ayahs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Ayah"
...
  /bookmarks:
    get:
      operationId: listBookmarks
      tags: [bookmarks]
      summary: Get all bookmarks
      responses:
        "200":
          description: Bookmarks list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Bookmark"
    post:
      operationId: createBookmark
      tags: [bookmarks]
      summary: Create a bookmark
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateBookmarkBody"
      responses:
        "201":
          description: Created bookmark
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Bookmark"
  /bookmarks/{bookmarkId}:
    delete:
      operationId: deleteBookmark
      tags: [bookmarks]
      summary: Delete a bookmark
      parameters:
        - name: bookmarkId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
components:
  schemas:
...
    Ayah:
      type: object
      properties:
        id:
          type: integer
        surahId:
          type: integer
        surahName:
          type: string
        ayahNumber:
          type: integer
        arabicText:
          type: string
        translation:
          type: string
        transliteration:
          type: string
        juzNumber:
          type: integer
        pageNumber:
          type: integer
        audioUrl:
          type: string
          nullable: true
      required: [id, surahId, surahName, ayahNumber, arabicText, translation, transliteration, juzNumber, pageNumber]
...
    Hadith:
      type: object
      properties:
        id:
          type: integer
        collectionId:
          type: string
        collectionName:
          type: string
        hadithNumber:
          type: string
        arabicText:
          type: string
        translation:
          type: string
        grade:
          type: string
          enum: [Sahih, Hasan, Da'if, Mawdu', Unknown]
        gradeReason:
          type: string
        narrator:
          type: string
        topics:
          type: array
          items:
            type: string
      required: [id, collectionId, collectionName, hadithNumber, arabicText, translation, grade, narrator, topics]
    HadithDetail:
      allOf:
        - $ref: "#/components/schemas/Hadith"
        - type: object
          properties:
            sharh:
              type: string
            isnadChain:
              $ref: "#/components/schemas/IsnadChain"
          required: []
    HadithListResponse:
      type: object
      properties:
        hadiths:
          type: array
          items:
            $ref: "#/components/schemas/Hadith"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required: [hadiths, total, page, limit]
    IsnadChain:
      type: object
      properties:
        hadithId:
          type: integer
        narrators:
          type: array
          items:
            $ref: "#/components/schemas/Narrator"
        overallGrade:
          type: string
      required: [hadithId, narrators, overallGrade]
    Narrator:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        nameArabic:
          type: string
        birthYear:
          type: integer
          nullable: true
        deathYear:
          type: integer
          nullable: true
        location:
          type: string
        reliability:
          type: string
        grade:
          type: string
        heardFrom:
          type: array
          items:
            type: string
        position:
          type: integer
      required: [id, name, nameArabic, reliability, grade, heardFrom, position]
...
    Bookmark:
      type: object
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [ayah, hadith]
        referenceId:
          type: integer
        title:
          type: string
        note:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required: [id, type, referenceId, title, createdAt]
    CreateBookmarkBody:
      type: object
      properties:
        type:
          type: string
          enum: [ayah, hadith]
        referenceId:
          type: integer
        title:
          type: string
        note:
          type: string
          nullable: true
      required: [type, referenceId, title]