# lib/api-zod/src/generated/types/bookmark.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Muhandis Islamic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { BookmarkType } from "./bookmarkType";

export interface Bookmark {
  id: number;
  type: BookmarkType;
  referenceId: number;
  title: string;
  note?: string | null;
  createdAt: Date;
}