# lib/api-zod/src/generated/types/createBookmarkBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Muhandis Islamic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { CreateBookmarkBodyType } from "./createBookmarkBodyType";

export interface CreateBookmarkBody {
  type: CreateBookmarkBodyType;
  referenceId: number;
  title: string;
  note?: string | null;
}