# lib/api-zod/src/generated/types/isnadChain.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Al-Muhandis Islamic Intelligence Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Narrator } from "./narrator";

export interface IsnadChain {
  hadithId: number;
  narrators: Narrator[];
  overallGrade: string;
}